Comparison between signed and unsigned integrals has been omitted.
git-svn-id: svn+ssh://cv077/home/ueshiba/svnroot/trunk/src@2449 3ef1b8b6-fb57-41bf-88d9-9a5bd220bf38

+++ TUv++/ListCmd.cc
@@ -113,7 +113,7 @@ ListCmd::setPercent(float percent)
     _top = int(_nitems * percent + 0.5);
     if (_top < 0)
 	_top = 0;
-    else if (_top > _nitems - nshown)
+    else if (size_t(_top) > _nitems - nshown)
 	_top = _nitems - nshown;
     vViewportSetLocation(_widget, 0.0, float(_top) / float(_nitems));
 }
